asdas

+++ project_back/controllers/team_member_controller.js
@@ -1,6 +1,6 @@
 import teamMemberService from "../services/team_member_service.js";
 import teamService from "../services/team_service.js";
-import  TeamMember  from "../models/TeamMember.js";
+import TeamMember from "../models/TeamMember.js";
 export const addMemberToTeam = async (req, res) => {
     try {
         const { teamId } = req.params;
@@ -40,6 +40,21 @@ export const requestToJoinTeam = async (req, res) => {
         const userId = req.user.id;
         const { teamId } = req.params;
 
+        const existingMember = await teamMemberService.getMemberOfTeam(teamId, userId);
+
+        if (existingMember) {
+            if (existingMember.status === "accepted") {
+                return res.status(400).json({ message: "Ya eres miembro de este equipo." });
+            }
+
+            if (existingMember.status === "pending") {
+                return res.status(400).json({ message: "Ya has solicitado unirte a este equipo." });
+            }
+
+
+            return res.status(400).json({ message: "No puedes volver a enviar la solicitud." });
+        }
+
         const teamMember = await teamMemberService.addMember({
             teamId,
             userId,
@@ -52,6 +67,7 @@ export const requestToJoinTeam = async (req, res) => {
         res.status(500).json({ message: error.message });
     }
 };
+
 
 export const rejectRequestToJoinTeam = async (req, res) => {
     try {
@@ -110,10 +126,10 @@ export const deleteMembersOfTeam = async (req, res) => {
         res.status(500).json({ message: error.message });
     }
 };
-
 export const deleteMemberOfTeam = async (req, res) => {
     try {
         const { teamId, userId } = req.params;
+        const requesterId = req.user.id;
 
         if (!userId || !teamId) {
             return res.status(400).json({ message: "Faltan datos" });
@@ -122,6 +138,11 @@ export const deleteMemberOfTeam = async (req, res) => {
         const member = await teamMemberService.getMemberOfTeam(teamId, userId);
         if (!member) {
             return res.status(404).json({ message: "El usuario no pertenece al equipo" });
+        }
+
+
+        if (member.role === "admin" && userId === requesterId) {
+            return res.status(403).json({ message: "No puedes eliminarte a ti mismo si eres administrador" });
         }
 
         await teamMemberService.removeMember({ teamId, userId });
@@ -133,14 +154,15 @@ export const deleteMemberOfTeam = async (req, res) => {
 };
 
 
+
 export const confirmJoinWithCode = async (req, res) => {
     try {
         const { teamId } = req.params;
         const { code } = req.body;
 
         console.log("teamId:", teamId);
-       
-        const verified = await teamService.getConfirmationCode(teamId );
+
+        const verified = await teamService.getConfirmationCode(teamId);
         if (!verified) return res.status(404).json({ message: "Verification code not found" });
 
         if (verified.code !== code) return res.status(400).json({ message: "Invalid code" });
@@ -150,26 +172,25 @@ export const confirmJoinWithCode = async (req, res) => {
         res.status(500).json({ message: error.message });
     }
 };
- 
+
 export const getPendingRequests = async (req, res) => {
-  try {
-    const userId = req.user.id;
-    const requests = await teamMemberService.getPendingRequests(userId);
-    res.status(200).json(requests.map(r => r.teamId));
-  } catch (error) {
-    res.status(500).json({ message: error.message });
-  }
+    try {
+        const userId = req.user.id;
+        const requests = await teamMemberService.getPendingRequests(userId);
+        res.status(200).json(requests.map(r => r.teamId));
+    } catch (error) {
+        res.status(500).json({ message: error.message });
+    }
 };
 
 export const getPendingMembersOfTeam = async (req, res) => {
-  try {
-    const { teamId } = req.params;
-    const pendingMembers = await teamMemberService.getPendingMembersOfTeam(teamId);
-    res.status(200).json(pendingMembers);
-  } catch (error) {
-    res.status(500).json({ message: error.message });
-  }
+    try {
+        const { teamId } = req.params;
+        const pendingMembers = await teamMemberService.getPendingMembersOfTeam(teamId);
+        res.status(200).json(pendingMembers);
+    } catch (error) {
+        res.status(500).json({ message: error.message });
+    }
 };
 
 
-
